Very strange. my CPU midpoint is no longer working well. the GPU is getting 1e-4 accuracy...dont know which one to believe.

+++ src/utils/opencl/cl_base_class.cpp
@@ -185,9 +185,9 @@ std::vector<std::string> CLBaseClass::split(const std::string &s, char delim) {
 // NOTE: if keep_substr=false we discard the delim wherever it is matched
 // if keep_substr=true we keep the delim in the substrings
 std::vector<std::string>& CLBaseClass::split(const std::string &s, const std::string delim, std::vector<std::string> &elems, bool keep_substr) {
-    unsigned int last_found = 0; 
-    unsigned int find_start = last_found; 
-    unsigned int found = 0;
+    size_t last_found = 0; 
+    size_t find_start = last_found; 
+    size_t found = 0;
     while(found != std::string::npos)
     {
         found = s.find(delim, find_start);
